updated to scalismo 0.91-RC2 (and vtk 9.1) (#57)
Changes required due to the update to vtk version 9.1

In previous versions, the vtkJoglPanelComponent was reimplemented and
extended. This caused problems on MacOS. It seems that the bugs that
were the reason that it was reimplemented in the first place are not
causing any problems (or are fixed?) we revert to the original
implementation.

+++ src/main/scala/scalismo/ui/rendering/actor/TransformationGlyphActor.scala
@@ -53,7 +53,7 @@ trait TransformationGlyphActor extends VectorFieldActor {
     for ((vector, _) <- sceneNode.transformedSource.values.zipWithIndex) {
       val norm = vector.norm
       vectors.InsertNextTuple3(vector(0), vector(1), vector(2))
-      scalars.InsertNextValue(norm)
+      scalars.InsertNextValue(norm.toFloat)
       if (norm > maxNorm) maxNorm = norm
       if (norm < minNorm) minNorm = norm
     }
